+++ src/main/java/de/slikey/effectlib/util/ParticleOptions.java
@@ -1,7 +1,9 @@
 package de.slikey.effectlib.util;
 
 import org.bukkit.Color;
+import org.bukkit.Location;
 import org.bukkit.Material;
+import org.bukkit.entity.Entity;
 
 public class ParticleOptions {
     public Color color;
@@ -11,10 +13,12 @@ public class ParticleOptions {
     public float offsetZ;
     public float speed;
     public int amount;
+    public int arrivalTime;
     public float size;
     public Material material;
     public byte materialData;
     public Object data;
+    public DynamicLocation target;
 
     public ParticleOptions() {
 
@@ -32,7 +36,7 @@ public class ParticleOptions {
         this.materialData = materialData;
     }
 
-    public ParticleOptions(float offsetX, float offsetY, float offsetZ, float speed, int amount, float size, Color color, Color toColor, Material material, byte materialData) {
+    public ParticleOptions(float offsetX, float offsetY, float offsetZ, float speed, int amount, float size, Color color, Color toColor, int arrivalTime, Material material, byte materialData) {
         this.offsetX = offsetX;
         this.offsetY = offsetY;
         this.offsetZ = offsetZ;
@@ -43,5 +47,6 @@ public class ParticleOptions {
         this.toColor = toColor;
         this.material = material;
         this.materialData = materialData;
+        this.arrivalTime = arrivalTime;
     }
 }